requirements.txt
alabaster==0.7.8
argh==0.26.2
astroid==1.4.5
Babel==2.3.4
colorama==0.3.7
docutils==0.12
httpsrv==0.1.5
imagesize==0.7.1
Jinja2==2.8
lazy-object-proxy==1.2.2
MarkupSafe==0.23
pathtools==0.1.2
Pygments==2.1.3
pylint==1.5.5
pytz==2016.4
PyYAML==3.11
requests==2.10.0
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.4.1
sphinx-rtd-theme==0.1.9
tornado==4.3
watchdog==0.8.3
wrapt==1.10.8
